feat: pluggable StateStore interface with SqliteStateStore backend (#176)
- Add StateStore interface (src/state/types.ts) with 34 methods + lifecycle
- Add SqliteStateStore class (src/state/sqlite-store.ts) implementing StateStore
  - All observability preserved: slow query warnings, classified error logging
  - Transactions on setChannelPrefs and addPermissionRule
- Refactor store.ts into thin facade delegating to active StateStore instance
 202 lines; all exports preserved for zero caller changes
  - Auto-initializes with SQLite defaults for backward compatibility
- Add database config option to AppConfig (database.module + options)
- Add plugin loading in index.ts: dynamic import of custom StateStore modules
- Update config.sample.json with database config example
